feat(service): 任务服务删除任务添加存在判断

+++ todo/src/main/java/org/snowzen/service/TaskService.java
@@ -222,6 +222,9 @@ public class TaskService {
     public void delete(int taskId) {
         checkArgument(IdUtil.checkId(taskId));
 
+        if (!taskRepository.existsById(taskId)) {
+            return;
+        }
         taskRepository.deleteById(taskId);
         tagTaskRelationRepository.deleteAllByTaskId(taskId);
         categoryTaskRelationRepository.deleteAllByTaskId(taskId);
